requirements.txt
Flask==3.0.0
qrcode==7.4.2
Pillow==10.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
colorama==0.4.6
pypng==0.20220715.0
